requirements.txt
streamlit==1.31.0
langchain==0.1.16
langchain-openai==0.0.8
langchain-community
faiss-cpu
pypdf==4.0.1
python-dotenv==1.0.1
openai==1.12.0
tiktoken==0.6.0
